+++ providers/kmpauth-google/src/wasmJsMain/kotlin/com/mmk/kmpauth/google/GoogleAuthUiProviderImpl.kt
@@ -22,7 +22,8 @@ internal class GoogleAuthUiProviderImpl(private val credentials: GoogleAuthCrede
     override suspend fun signIn(
         filterByAuthorizedAccounts: Boolean,
         isAutoSelectEnabled: Boolean,
-        scopes: List<String>
+        scopes: List<String>,
+        requestAccessToken: Boolean
     ): Result<GoogleUser> {
 
         val scriptLoaded = waitForGoogleAuthScriptToLoad()